x AND NOTy OR y AND z is ready #boolean #logic #circuits #simulator

+++ proj173num5/Value.c
@@ -32,5 +32,5 @@ void Value_setValue(Value* this, bool b) {
 }
 
 void Value_print(Value* this) {
-	printf("0x%lx=%s", this, this->value ? "true" : "false");
+    printf("0x%lx=%s", (unsigned long)this, this->value ? "true" : "false");
 }
